Dispatch test_bugs into each own VTherm type tests
